requirements.txt
discord~=2.3.2
python-dotenv~=1.1.0
requests~=2.32.3
beautifulsoup4~=4.13.3
numpy~=2.2.4
pillow~=11.2.0